refactor: move transaction types to a new module and update imports
